consolidate serializers
