+++ fcrepo-event-serialization/src/main/java/org/fcrepo/event/serialization/TurtleSerializer.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.fcrepo.event.serialization.rdf;
+package org.fcrepo.event.serialization;
 
 import java.io.ByteArrayOutputStream;
 import java.io.UnsupportedEncodingException;
